doc xlc optimization bug (reported to IBM)
(x & 0) is always 0, and (x ^ 0) is always x
xlc optimizes away the load of x when an operand to '&' or '^' is
constant 0 and the value of x is previously known.  This is fine
unless load of x is load-and-reserve instruction (e.g. lwarx or ldarx)
required before a subsequent store-conditional (e.g. stwcx. or stdcx.)